backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
oracledb==2.4.1
langgraph==0.1.18
langgraph-checkpoint-sqlite==0.1.3
langchain==0.2.10
langchain-core==0.2.16
langchain-openai==0.1.7
faiss-cpu==1.8.0
chromadb==0.5.0
tavily-python==0.3.6
